Create taskUtils.js with functions and constants for tasks
